// frontend/src/api/generated/types/wordDetail.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Language Learning API
 * Interactive API documentation for the auth, profile, and words endpoints.
 * OpenAPI spec version: 1.0.0
 */
import type { WordLanguage } from './wordLanguage';
import type { WordMeaning } from './wordMeaning';

export interface WordDetail {
  id: number;
  value: string;
  /** @nullable */
  frequencyRank: number | null;
  language: WordLanguage;
  meanings: WordMeaning[];
}